+++ src/index.js
@@ -1,43 +1,40 @@
 import './style.css';
-import Task from './task.js';
-import LocalStorage from './localstorage.js';
+import {Task, getData, saveData} from './task.js';
 import UI from './UI.js';
+import display from './display.js';
 
 let tasksList;
-if (LocalStorage.getData() === null) {
+if (getData() === null) {
   tasksList = [];
 } else {
-  tasksList = LocalStorage.getData();
+  tasksList = getData();
 }
 
 const addTask = (newTask) => {
   let index;
-  if (LocalStorage.getData() === null) {
+  if (getData() === null) {
     index = 1;
   } else {
-    tasksList = LocalStorage.getData();
+    tasksList = getData();
     index = tasksList.length + 1;
   }
   const task = new Task(newTask, false, index);
   tasksList.push(task);
-  LocalStorage.saveData(tasksList);
-  UI.showAllTasks(tasksList);
+  saveData(tasksList);
+  display(tasksList);
 };
 
-const clearInput = () => {
-  document.querySelector('#add-new-task').value = '';
-};
 
 const addNewTask = document.querySelector('#add-new-task');
 addNewTask.addEventListener('keyup', (e) => {
   if (e.keyCode === 13 && addNewTask.value !== '') {
     const newTask = addNewTask.value;
     addTask(newTask);
-    clearInput();
+    document.querySelector('#add-new-task').value = '';
   }
 });
 
-UI.showAllTasks(tasksList);
+display(tasksList);
 
 const btnRefresh = document.querySelector('#btn-refresh');
 btnRefresh.addEventListener('click', () => {
