+++ src/ui.js
@@ -1,4 +1,5 @@
-import LocalStorage from './localstorage.js';
+import {getData, saveData} from './task.js';
+import display from './display.js';
 
 export default class UI {
   static updateIndex(tasksList) {
@@ -12,46 +13,26 @@ export default class UI {
     const currentLi = target.parentElement;
     currentLi.parentElement.removeChild(currentLi);
 
-    let tasksList = LocalStorage.getData();
+    let tasksList = getData();
     tasksList = tasksList.filter((task, idx) => idx !== index);
     tasksList = this.updateIndex(tasksList);
-    LocalStorage.saveData(tasksList);
-    this.showAllTasks(tasksList);
+    saveData(tasksList);
+    display(tasksList);
   }
 
   static editTask(value, index, tasksList) {
     tasksList[index].description = value;
-    LocalStorage.saveData(tasksList);
-    this.showAllTasks(tasksList);
+    saveData(tasksList);
+    display(tasksList);
   }
 
   static reloadPage() {
-    const tasksList = LocalStorage.getData();
+    const tasksList = getData();
     tasksList.forEach((task) => {
       task.completed = false;
     });
-    LocalStorage.saveData(tasksList);
+    saveData(tasksList);
   }
 
-  static showAllTasks(tasksList) {
-    const todoList = document.querySelector('.todo-list');
-    let tasks = '';
 
-    tasksList.forEach((task) => {
-      tasks += `<li class="todo-item">
-      <input type="checkbox" name="check" id="check">
-      <input type="text" name="task" id="task" value="${task.description}"" reuired>
-      <i class="fa-solid fa-trash-can btn-delete"></i>
-      </li>`;
-    });
-
-    todoList.innerHTML = tasks;
-
-    const deleteBtn = document.querySelectorAll('.btn-delete');
-    deleteBtn.forEach((btn, index) => {
-      btn.addEventListener('click', (e) => {
-        this.removeTask(e.target, btn, index);
-      });
-    });
-  }
 }
